Rename metaclasses to metaclass (#36)
