+++ model/src/pyrenew/observation/negativebinomial.py
@@ -5,7 +5,7 @@ import numbers as nums
 import numpyro
 import numpyro.distributions as dist
 from numpy.typing import ArrayLike
-from pyrenew.metaclasses import RandomVariable
+from pyrenew.metaclass import RandomVariable
 
 
 class NegativeBinomialObservation(RandomVariable):
