+++ model/src/pyrenew/process/simplerandomwalk.py
@@ -3,7 +3,7 @@
 import jax.numpy as jnp
 import numpyro as npro
 import numpyro.distributions as dist
-from pyrenew.metaclasses import RandomVariable
+from pyrenew.metaclass import RandomVariable
 
 
 class SimpleRandomWalkProcess(RandomVariable):
